Refactored all basic web services into Python class
Websocket service and client webbrowser start-up hidden
behind an Python class that can be overriden for
different types of functionalities.  Still need to
add the option to start for a web server to serve the UI files,
in addition to reading directly off of the file system.
The former is required for remote web-based UI; the latter
is sufficient for local applications.